+++ common/src/main/kotlin/com/noheltcj/rxcommon/observers/NextObserver.kt
@@ -1,5 +1,3 @@
 package com.noheltcj.rxcommon.observers
 
-import com.noheltcj.rxcommon.observable.AbstractObserver
-
 class NextObserver<E>(next: (E) -> Unit) : AbstractObserver<E>(next = next)
