+++ common/src/main/kotlin/com/noheltcj/rxcommon/subjects/PublishSubject.kt
@@ -1,6 +1,6 @@
 package com.noheltcj.rxcommon.subjects
 
-import com.noheltcj.rxcommon.observable.ObservableEmitter
+import com.noheltcj.rxcommon.emitters.ObservableEmitter
 
 class PublishSubject<E> : Subject<E>() {
   override val emitter = ObservableEmitter<E>()
